backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg[binary]==3.2.1
celery==5.4.0
redis==5.0.8
gunicorn==22.0.0
